remove steering test file
