+++ Assets/Scripts/NPC/StoppingCondition.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 25e686b48a5f1db4f8c6621aa9054757
-timeCreated: 1454798768
+guid: d797ee10194474b40bbe5f8fc0e3230e
+timeCreated: 1456201891
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
